drm/amdgpu: Remove immutable flag from freesync_capable property
Atomic drivers should not use legacy properties but atomic properties
don't support the immutable flag. Remove the immutable flag for now.

This will be followed by a bunch of DC patches to start treating
freesync_capable as an atomic property on a connector_state.

Eventually we'll want to remove the these properties and replace them
with new atomic adaptive sync properties that we define in DRM.

Signed-off-by: Harry Wentland <harry.wentland@amd.com>
Reviewed-by: Andrey Grodzovsky <andrey.grodzovsky@amd.com>
Signed-off-by: Alex Deucher <alexander.deucher@amd.com>