UBUNTU: [Config] CONFIG_UNWINDER_FRAME_POINTER=y for amd64
During the rebase to 4.15 UNWINDER_ORC was chosen as the stack
unwinder as it promises a 5-10% performance improvement over
using UNWINDER_FRAME_POINTER. However it turns out to have a
couple of downsides. It adds a new requierment for building dkms
modules, and it does not produce the reliable stack traces
required for livepatch. Switch back to UNWINDER_FRAME_POINTER.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>