ASoC: cs35l41: Remove DAPM route for ASP TX zero-fill
Remove DAPM routes for ASP TX widgets when zero-fill
is selected as the source. "Zero" should be selected
for the ASP TX source whenever that channel is inactive.

For example, when capturing one channel of data from
each amp in a stereo setup, ASP TX2, TX3, and TX4 should
all have source set to "Zero".

Signed-off-by: David Rhodes <david.rhodes@cirrus.com>