UBUNTU: SAUCE: IB/mlx5: use size_t instead of u64 when dividing
Currently several length values used in division operations are
64-bits, which can cause references to library helpers not
available in the kernel. For example, on i386:

 ERROR: "__udivdi3" [drivers/infiniband/hw/mlx5/mlx5_ib.ko] undefined!

And on armhf:

 ERROR: "__aeabi_uldivmod" [drivers/infiniband/hw/mlx5/mlx5_ib.ko] undefined!

In some cases the vale being passed to a u64 function argument is
always from a size_t variable, and it's clearly safe to switch to
a size_t for these cases. In the other case the value is in
practice limited to the use of an unsigned long due to the return
type of roundup_pow_of_two(), but go ahead and add an extra check
for overflow to be on the safe side.

This issue has been reported upstream but no patch is yet
available. This patch should be dropped once a patch is upstream.

Signed-off-by: Seth Forshee <seth.forshee@canonical.com>