UBUNTU: [Config] Enable notifier call chain validations
BugLink: https://launchpad.net/bugs/1855337

Enable CONFIG_DEBUG_NOTIFIERS to ensure that notifier functions are
present in the core kernel text or module text sections before calling
those functions.

If an invalid function pointer is detected, a warning is issued and the
function is not called. This helps in attack prevention and detection.

Signed-off-by: Tyler Hicks <tyhicks@canonical.com>
Signed-off-by: Seth Forshee <seth.forshee@canonical.com>