ASoC: cs35l41: Fix PDN failed message
The driver could report an erroneous 'PDN failed' message on first
boot.  This is because the PDN_DONE is interrupt driven, and DAPM
may trigger the event handler to power down the chip when it is
already powered down. Thus, the monitored PDN_DONE interrupt never
fires and the driver will report a false 'PDN failed' message.

This patch ensures that the power down routine only executes when the
amp is not already powered down.

Signed-off-by: Adam Brickman <Adam.Brickman@cirrus.com>