Add support for AMD SPI controller-1
[Why]
SPI controller is connected to x86. The design of controller is
such that for  every transfer it sends the opcode. Currently for
tx opcode as 0x0b and rx as 0x0a is fixed

[How]
Add ctrl_id for AMDI0062 SPI controller.
Add support for SPI based TPS Client
remove rx/tx count set

Signed-off-by: Nehal Bakulchandra Shah <nehal-bakulchandra.shah@amd.com>
Signed-off-by: Kun Liu <Kun.Liu2@amd.com>